requirements.txt
Flask==3.0.3
python-dotenv==1.0.1
requests==2.32.3
pydantic>=2.0
redis==5.0.8
gunicorn==22.0.0
werkzeug==3.0.3

# Dev & QA
pytest==8.3.3
pytest-flask==1.3.0
responses==0.25.3
pre-commit==3.8.0
black==24.8.0
isort==5.13.2
flake8==7.1.1
